requirements.txt
flask
discord
google-generativeai
pyngrok
Pillow
python-dotenv